+++ tests/test_wsgidav_app.py
@@ -59,7 +59,7 @@ class ServerTest(unittest.TestCase):
             "property_manager": None,  # None: no property manager
             "lock_manager": True,  # True: use lock_manager.LockManager
             # None: domain_controller.SimpleDomainController(user_mapping)
-            "domain_controller": None,
+            # "domain_controller": None,
         }
 
         if withAuthentication: