+++ wsgidav/default_conf.py
@@ -14,7 +14,7 @@
 
 Default confguration.
 """
-from wsgidav.addons.dir_browser import WsgiDavDirBrowser
+from wsgidav.dir_browser import WsgiDavDirBrowser
 from wsgidav.debug_filter import WsgiDavDebugFilter
 from wsgidav.error_printer import ErrorPrinter
 from wsgidav.http_authenticator import HTTPAuthenticator